# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "golang作品",
        "title": "Gin-Backend",
        "termsOfService": "github",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/order/add": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Order OrderInfo",
                "parameters": [
                    {
                        "description": "Add Order",
                        "name": "o",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "請求錯誤",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "內部錯誤",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/order/delete": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Order EditOrder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Delete Order",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "請求錯誤",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "內部錯誤",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/order/edit": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Order EditOrder",
                "parameters": [
                    {
                        "description": "Edit Order",
                        "name": "o",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "請求錯誤",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "內部錯誤",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/order/info": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Order OrderInfo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order Info",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "請求錯誤",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "內部錯誤",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/order/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Order OrderInfo",
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "請求錯誤",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "內部錯誤",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Order": {
            "type": "object",
            "properties": {
                "createAt": {
                    "type": "string"
                },
                "extraInfo": {
                    "type": "string"
                },
                "isDeleted": {
                    "type": "boolean"
                },
                "mobile": {
                    "type": "string"
                },
                "nickName": {
                    "type": "string"
                },
                "orderDetail": {
                    "$ref": "#/definitions/model.OrderDetailItems"
                },
                "orderId": {
                    "type": "string"
                },
                "orderStatus": {
                    "type": "integer"
                },
                "payStatus": {
                    "type": "integer"
                },
                "payTime": {
                    "type": "string"
                },
                "payType": {
                    "type": "integer"
                },
                "totalPrice": {
                    "description": "gorm:\"column:order_detail\"",
                    "type": "number"
                },
                "updateAt": {
                    "type": "string"
                },
                "userAddress": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "model.OrderDetailItems": {
            "type": "object",
            "properties": {
                "itemCount": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "itemId": {
                    "description": "id price",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    }
}